+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/afk/AfkInitializer.java
@@ -9,7 +9,6 @@ import io.github.sakurawald.fuji.core.command.annotation.CommandSource;
 import io.github.sakurawald.fuji.core.command.annotation.CommandTarget;
 import io.github.sakurawald.fuji.core.config.handler.abst.BaseConfigurationHandler;
 import io.github.sakurawald.fuji.core.config.handler.impl.ObjectConfigurationHandler;
-import io.github.sakurawald.fuji.core.extension.PlayerCombatExtension;
 import io.github.sakurawald.fuji.core.document.annotation.ColorBox;
 import io.github.sakurawald.fuji.module.initializer.ModuleInitializer;
 import io.github.sakurawald.fuji.module.initializer.afk.config.model.AfkConfigModel;
@@ -42,11 +41,7 @@ public class AfkInitializer extends ModuleInitializer {
     @Document(id = 1751826266551L, value = "Enter afk state.")
     private static int $afk(@CommandSource @CommandTarget ServerPlayerEntity player) {
         // NOTE: Issue a command will update the lastLastActionTime, so it's impossible to use /afk to disable afk
-        if (!player.isOnGround()
-            || player.isOnFire()
-            || player.inPowderSnow
-            || ((PlayerCombatExtension) player).fuji$inCombat()) {
-
+        if (AfkService.canAfk(player)) {
             TextHelper.sendTextByKey(player, "afk.on.failed");
             return CommandHelper.Return.FAILURE;
         }
@@ -57,9 +52,9 @@ public class AfkInitializer extends ModuleInitializer {
     }
 
     @Document(id = 1751826271499L, value = "Test if a player is in afk state.")
-    @CommandNode("test-afk")
+    @CommandNode("is-afk?")
     @CommandRequirement(level = 4)
-    private static int $testAfk(@CommandSource ServerCommandSource source, ServerPlayerEntity player) {
+    private static int $isAfk(@CommandSource ServerCommandSource source, ServerPlayerEntity player) {
         boolean value = AfkService.isAfk(player);
         return CommandHelper.Return.returnBoolean(source, value);
     }